% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_FunctPFOpt.R
\name{BLPortfolioOptimizer}
\alias{BLPortfolioOptimizer}
\title{BLPortfolioOptimizer}
\usage{
BLPortfolioOptimizer(risk.aversion.coeff = 3.07, tau = 0.025,
  covar.matrix, market.cap.weights, ident.view.matrix = NA,
  diag.covar.error.matrix = NA, view.vector = NA)
}
\arguments{
\item{risk.aversion.coeff}{(lambda) risk aversion coefficient, default is 3}

\item{tau}{default is 0.025}

\item{covar.matrix}{Covariance Matrix of N assets (N x N Matrix)}

\item{market.cap.weights}{is a vector with the market capitalization weights}

\item{ident.view.matrix}{(BL: P) Matrix that identifies the assets involved in the views (K x N Matrix)}

\item{diag.covar.error.matrix}{(BL: Omega)
A diagonal covariance matrix of error terms from the expressed views
 representing the uncertainty in each view ( K x K Matrix)}

\item{view.vector}{(BL: Q) Vector including the Views (K x 1 column vector)}
}
\value{
cvar CVaR of the specific portfolio or asset with the set alpha
}
\description{
Black Litterman Portfolio Optimizer
}
