% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_FunctPFOpt.R
\name{CVaRPortfolioOptimizer}
\alias{CVaRPortfolioOptimizer}
\title{CVaRPortfolioOptimizer}
\usage{
CVaRPortfolioOptimizer(daily.returns.data.wide, alpha.cvar = 0.05,
  rmin = 0, wmin = 0, wmax = 1, weight.sum = 1)
}
\arguments{
\item{daily.returns.data.wide}{data.frame including the daily returns of the specific assets as well as a reference date}

\item{rmin}{Required return, default is 0}

\item{wmin}{Minimal weight of a single asset, default is 0}

\item{wmax}{Maximal weight of a single asset, default is 1}

\item{weight.sum}{Total weight of all assets, default is 1}

\item{alpha}{Alpha of the CVaR, this is the confidence level from which on the average of the tail risk is being calculated}
}
\value{
cvar CVaR of the specific portfolio or asset with the set alpha
}
\description{
CVaR Portfolio Optimizer based on Yollin (2009)
}
