% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{generateOptimizationResultStats}
\alias{generateOptimizationResultStats}
\title{generateOptimizationResultStats}
\usage{
generateOptimizationResultStats(out.of.sample.period.months = 24,
  investment.period.months = 12, daily.returns.data.wide,
  pf.opt.type = "mpt.min.var", use.covar.matrix = FALSE,
  covar.type = "sample", correl.type = "sample",
  vola.type = "sample", exp.return = "sample", target.return = NA,
  sum.weight = 1, min.single.weight = -100, max.single.weight = 100,
  alpha.cvar = NA, bl.risk.aversion.coeff = 3.07, bl.tau = 0.025,
  bl.market.cap.weights = NA, bl.ident.view.matrix = NA,
  bl.diag.covar.error.matrix = NA, bl.view.vector = NA,
  in.sample = FALSE)
}
\arguments{
\item{out.of.sample.period.months}{how many months are considered for the out of sample period}

\item{investment.period.months}{number of months of the investment horizon before the next rebalancing happens}

\item{daily.returns.data.wide}{data frame with returns with out of sample return data as well as data for the actual investment period}

\item{pf.opt.type}{type of optimization, min.var,}

\item{covar.type}{either sample of}

\item{correl.type}{either sample of}

\item{vola.type}{sample}

\item{exp.return}{expected return of the assets, default is sample - meaning that we take historical asset return as a best predictor}

\item{target.return}{default is NA}

\item{sum.weight}{default is 1}

\item{min.single.weight}{default is -100}

\item{max.single.weight}{default is +100}

\item{alpha.cvar}{Alpha parameter for CVaR optimization only, default is NA}

\item{in.sample}{default is FALSE}
}
\value{
list with result data.frames: weights.result.table,
}
\description{
generate Optimization Result Stats
}
