% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{histCVaRcalc}
\alias{histCVaRcalc}
\title{histCVaRcalc}
\usage{
histCVaRcalc(asset.weights = 1, daily.returns.data.wide, alpha.cvar)
}
\arguments{
\item{alpha.cvar}{Alpha of the CVaR, this is the confidence level from which on the average of the tail risk is being calculated}

\item{daily.returns}{Daily historical returns of a portfolio or single asset}
}
\value{
cvar CVaR of the specific portfolio or asset with the set alpha
}
\description{
calculates the historical Conditional VaR / Expected Shortfall of a portfolio or an asset
}
