% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{meanVariancePortfolioOptimizer}
\alias{meanVariancePortfolioOptimizer}
\title{meanVariancePortfolioOptimizer}
\usage{
meanVariancePortfolioOptimizer(asset.name, mu.vector, sigma.vector,
  correl.matrix, target.return, rf, print.out = FALSE)
}
\arguments{
\item{mu.vector}{vector with estimated returns of investment objects}

\item{sigma.vector}{vector with the volatilities of the investment objects}

\item{correl.matrix}{correlation matrix of the investment objects}

\item{target.return}{which return level is seeked (for which the variance is minimized)}

\item{rf}{risk free return}
}
\value{
weight.risky.assets a vector with the weights of the risky assets
}
\description{
This function creates all important descriptive statistics such as VaR, ES, Return for a portfolio
}
\examples{
weights.vector          <- c(0.7,0.3)
daily.returns.data.wide <- data.frame(ref.date=c(Sys.Date()-2:0), asset1.ret=c(-0.02,0.005,0.004), asset2.ret=c(0,-0.001,0.02))
PFstats(weights.vector=weights.vector, daily.returns.data.wide=daily.returns.data.wide)
}
