% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_FunctPFOpt.R
\name{meanVariancePortfolioOptimizerQP}
\alias{meanVariancePortfolioOptimizerQP}
\title{meanVariancePortfolioOptimizerQP}
\usage{
meanVariancePortfolioOptimizerQP(mu.vector, sigma.vector = NA,
  correl.matrix = NA, covar.matrix = NA, use.covar.matrix = FALSE,
  target.return, sum.weight = 1, min.single.weight = -100,
  max.single.weight = 100, mvp = FALSE)
}
\arguments{
\item{mu.vector}{vector with estimated returns of investment objects}

\item{sigma.vector}{vector with the volatilities of the investment objects}

\item{correl.matrix}{correlation matrix of the investment objects}

\item{use.covar.matrix}{use covariance matrix directly or build it via sigma vector and the correlation matrix, default is FALSE}

\item{target.return}{which return level is seeked (for which the variance is minimized)}

\item{sum.weight}{default is 1}

\item{min.single.weight}{default is -100}

\item{max.single.weight}{default is +100}

\item{asset.name}{Name of assets available}

\item{covar.matrx}{covariance matrix}
}
\value{
weight.risky.assets a vector with the weights of the risky assets
}
\description{
Mean Variance Portfolio Optimizer using Quadratic Programming
}
\examples{
weights.vector          <- c(0.7,0.3)
daily.returns.data.wide <- data.frame(ref.date=c(Sys.Date()-2:0), asset1.ret=c(-0.02,0.005,0.004), asset2.ret=c(0,-0.001,0.02))
PFstats(weights.vector=weights.vector, daily.returns.data.wide=daily.returns.data.wide)
}
