% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{PFstats}
\alias{PFstats}
\title{PFstats}
\usage{
PFstats(weights.vector, daily.returns.data.wide)
}
\arguments{
\item{weights.vector}{vector that contains the relative weights of the individual assets of the portfolio}

\item{daily.returns.data.wide}{data.frame including the daily returns of the specific assets as well as a reference date}
}
\value{
a list with a data.frame PF.return.result.table that includes all statistics, as well as two graphs - histogram with returns and a correlation matrix
}
\description{
This function creates all important descriptive statistics such as VaR, ES, Return for a portfolio
}
\examples{
weights.vector          <- c(0.7,0.3)
daily.returns.data.wide <- data.frame(ref.date=c(Sys.Date()-2:0), asset1.ret=c(-0.02,0.005,0.004), asset2.ret=c(0,-0.001,0.02))
PFstats(weights.vector=weights.vector, daily.returns.data.wide=daily.returns.data.wide)
}
