% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{SingleTitlestats}
\alias{SingleTitlestats}
\title{SingleTitlestats}
\usage{
SingleTitlestats(daily.returns.data.wide, num.trade.days.per.year = 250)
}
\arguments{
\item{daily.returns.data.wide}{data.frame including the daily returns of the specific assets as well as a reference date}

\item{num.trade.days.per.year}{Number of trading days per year, default set to 250}
}
\value{
a list with a data.frame PF.return.result.table that includes all statistics, as well as two graphs - histogram with returns and a correlation matrix
}
\description{
This function creates Single Title stats
}
