% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{generateOptimizationResultStats}
\alias{generateOptimizationResultStats}
\title{generateOptimizationResultStats}
\usage{
generateOptimizationResultStats(out.of.sample.period.months = 24,
  investment.period.months = 12, daily.returns.data.wide,
  pf.opt.type = "mpt.min.var", covar.type = "sample",
  correl.type = "sample", vola.type = "sample",
  exp.return = "sample", target.return = NA)
}
\arguments{
\item{out.of.sample.period.months}{how many months are considered for the out of sample period}

\item{investment.period.months}{number of months of the investment horizon before the next rebalancing happens}

\item{daily.returns.data.wide}{data frame with returns with out of sample return data as well as data for the actual investment period}

\item{pf.opt.type}{type of optimization, min.var,}

\item{covar.type}{either sample of}

\item{correl.type}{either sample of}

\item{vola.type}{sample}

\item{exp.return}{expected return of the assets, default is sample - meaning that we take historical asset return as a best predictor}

\item{target.return}{default is NA}
}
\value{
list with result data.frames: weights.result.table,
}
\description{
generate Optimization Result Stats
}
