% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQF_FA_Functions.R
\name{meanVariancePortfolioOptimizer}
\alias{meanVariancePortfolioOptimizer}
\title{meanVariancePortfolioOptimizer}
\usage{
meanVariancePortfolioOptimizer(asset.name, mu.vector, sigma.vector,
  correl.matrix, covar.matrix = NA, use.covar.matrix = FALSE,
  target.return, rf, print.out = FALSE, opt.focus.type = "return")
}
\arguments{
\item{asset.name}{Name of assets available}

\item{mu.vector}{vector with estimated returns of investment objects}

\item{sigma.vector}{vector with the volatilities of the investment objects}

\item{correl.matrix}{correlation matrix of the investment objects}

\item{use.covar.matrix}{use covariance matrix directly or build it via sigma vector and the correlation matrix, default is FALSE}

\item{target.return}{which return level is seeked (for which the variance is minimized)}

\item{rf}{risk free return}

\item{covar.matrx}{covariance matrix}
}
\value{
weight.risky.assets a vector with the weights of the risky assets
}
\description{
MPT mean Variance Portfolio Optimizer
}
\examples{
weights.vector          <- c(0.7,0.3)
daily.returns.data.wide <- data.frame(ref.date=c(Sys.Date()-2:0), asset1.ret=c(-0.02,0.005,0.004), asset2.ret=c(0,-0.001,0.02))
PFstats(weights.vector=weights.vector, daily.returns.data.wide=daily.returns.data.wide)
}
